Update ouroboros-consensus and update to cardano-api-8.8.0.0 and cardano-cli-8.3.0.0
Set initial block forging

Moved UseLedger type to avoid circular imports

- The moving of UseLedger was needed in order to import `Cardano.Node.Protocol (ProtocolInstantiationError)`
- Added new constructors to Startup Tracer

Update block forging configuration on SIGHUP

Block Forging can be enabled/disabled through SIGHUP signal. Sending
such a signal will trigger the node to read the block forging credential
files. Since the credential files are passed via CLI flags one can not
remove them without restarting the node. For this effect, in order for
one to be able to disable block forging, moving/renaming/deleting the
file at the specified path (for the credential flags) and then sending
the SIGHUP signal is what needs to be done. The code will detect that
the specified files do not exist and it will disable block forging,
while tracing the appropriate log messages.

In case one wants to start a block producing node (i.e. passing the
credentials in the respective flags) but doesn't want it to actually
behave as a BP, there's a new `--start-as-non-producing-node` flag that
will run the node with credentials as a normal node. However on SIGHUP
it will read the credential files and start forging.

Refactor code to accomodate the changes

Added EnabledBlockForging type

Fix Shutdown test for new exit codes

Co-authored-by: Armando Santos <armandoifsantos@gmail.com>
Co-authored-by: Jordan Millar <jordan.millar@iohk.io>
Co-authored-by: Marcin Szamotulski <coot@coot.me>
Co-authored-by: Samuel Leathers <samuel.leathers@iohk.io>